requirements.txt
requests==2.21.0
beautifulsoup4==4.8.2
argparse==1.4.0
pyyaml==5.3
